wip: add shared library
