+++ src/daemon/main.rs
@@ -1,6 +1,8 @@
 #![allow(clippy::cargo_common_metadata)]
 #![allow(clippy::multiple_crate_versions)]
 
+mod blctld;
+
 fn main() {
     println!("Daemon init");
 }
