+++ src/shared/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
-name = "blctl"
+name = "blctl-shared"
 author = "xela"
-version = "0.3.1"
+version = "0.1.0"
 edition = "2021"
 license = "MIT OR Apache-2.0"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
@@ -10,12 +10,6 @@ license = "MIT OR Apache-2.0"
 clap = { version = "4.5.16", features = ["derive"] }
 bincode = "1.3.3"
 
-blctl-shared = { path = "src/shared" }
-
-[[bin]]
-name = "blctl"
-path = "src/client/main.rs"
-
-[[bin]]
-name = "blctld"
-path = "src/daemon/main.rs"
+[lib]
+name = "blctl_shared"
+path = "src/lib.rs"
